Refactor local copy module structure
