+++ crates/engine/src/local_copy/dir_merge/mod.rs
@@ -1,0 +1,8 @@
+mod load;
+mod parse;
+
+pub(crate) use load::{
+    apply_dir_merge_rule_defaults, filter_program_local_error, load_dir_merge_rules_recursive,
+    resolve_dir_merge_path,
+};
+pub(crate) use parse::{FilterParseError, ParsedFilterDirective, parse_filter_directive_line};
