+++ crates/engine/src/local_copy/mod.rs
@@ -49,6 +49,8 @@
 //! assert_eq!(summary.files_copied(), 1);
 //! ```
 
+mod dir_merge;
+mod error;
 mod filter_program;
 mod options;
 mod overrides;
@@ -62,24 +64,30 @@ pub use plan::{
 
 pub use options::{DeleteTiming, LocalCopyOptions, ReferenceDirectory, ReferenceDirectoryKind};
 
+pub use error::{LocalCopyArgumentError, LocalCopyError, LocalCopyErrorKind};
+
 #[cfg(test)]
 pub(crate) use plan::FilterOutcome;
 
 pub use skip_compress::{SkipCompressList, SkipCompressParseError};
 
+#[allow(unused_imports)]
+pub(crate) use dir_merge::{
+    FilterParseError, ParsedFilterDirective, apply_dir_merge_rule_defaults,
+    filter_program_local_error, load_dir_merge_rules_recursive, parse_filter_directive_line,
+    resolve_dir_merge_path,
+};
+
 use self::filter_program::{
-    DirMergeParser, ExcludeIfPresentLayers, ExcludeIfPresentStack, FilterContext, FilterSegment,
-    FilterSegmentLayers, FilterSegmentStack, INVALID_OPERAND_EXIT_CODE, MAX_DELETE_EXIT_CODE,
-    MISSING_OPERANDS_EXIT_CODE, TIMEOUT_EXIT_CODE, directory_has_marker,
+    ExcludeIfPresentLayers, ExcludeIfPresentStack, FilterContext, FilterSegment,
+    FilterSegmentLayers, FilterSegmentStack, directory_has_marker,
 };
 use std::cell::{Cell, RefCell};
 use std::cmp::Ordering;
 #[cfg(unix)]
 use std::collections::HashMap;
 use std::collections::{HashSet, VecDeque};
-use std::error::Error;
 use std::ffi::{OsStr, OsString};
-use std::fmt;
 use std::fs;
 use std::io::{self, Read, Seek, SeekFrom, Write};
 use std::num::NonZeroU8;
@@ -1805,860 +1813,6 @@ impl Drop for DirectoryFilterGuard {
     }
 }
 
-fn filter_program_local_error(path: &Path, error: FilterProgramError) -> LocalCopyError {
-    LocalCopyError::io(
-        "compile filter file",
-        path.to_path_buf(),
-        io::Error::new(io::ErrorKind::InvalidData, error.to_string()),
-    )
-}
-
-fn resolve_dir_merge_path(base: &Path, pattern: &Path) -> PathBuf {
-    if pattern.is_absolute() {
-        if let Ok(stripped) = pattern.strip_prefix(Path::new("/")) {
-            return base.join(stripped);
-        }
-    }
-
-    base.join(pattern)
-}
-
-fn apply_dir_merge_rule_defaults(mut rule: FilterRule, options: &DirMergeOptions) -> FilterRule {
-    if options.anchor_root_enabled() {
-        rule = rule.anchor_to_root();
-    }
-
-    if let Some(sender) = options.sender_side_override() {
-        rule = rule.with_sender(sender);
-    }
-
-    if let Some(receiver) = options.receiver_side_override() {
-        rule = rule.with_receiver(receiver);
-    }
-
-    rule
-}
-
-#[derive(Default)]
-struct DirMergeEntries {
-    rules: Vec<FilterRule>,
-    exclude_if_present: Vec<ExcludeIfPresentRule>,
-}
-
-impl DirMergeEntries {
-    fn push_rule(&mut self, rule: FilterRule) {
-        self.rules.push(rule);
-    }
-
-    fn push_exclude_if_present(&mut self, rule: ExcludeIfPresentRule) {
-        self.exclude_if_present.push(rule);
-    }
-
-    fn extend(&mut self, mut other: DirMergeEntries) {
-        self.rules.append(&mut other.rules);
-        self.exclude_if_present
-            .append(&mut other.exclude_if_present);
-    }
-}
-
-fn load_dir_merge_rules_recursive(
-    path: &Path,
-    options: &DirMergeOptions,
-    visited: &mut Vec<PathBuf>,
-) -> Result<DirMergeEntries, LocalCopyError> {
-    let canonical = fs::canonicalize(path).unwrap_or_else(|_| path.to_path_buf());
-    if visited.contains(&canonical) {
-        let message = format!("recursive filter merge detected for '{}'", path.display());
-        return Err(LocalCopyError::io(
-            "parse filter file",
-            path.to_path_buf(),
-            io::Error::new(io::ErrorKind::InvalidData, message),
-        ));
-    }
-
-    visited.push(canonical);
-
-    let file = fs::File::open(path)
-        .map_err(|error| LocalCopyError::io("read filter file", path.to_path_buf(), error))?;
-    let mut entries = DirMergeEntries::default();
-
-    let map_error = |error: FilterParseError| {
-        LocalCopyError::io(
-            "parse filter file",
-            path.to_path_buf(),
-            io::Error::new(io::ErrorKind::InvalidData, error.to_string()),
-        )
-    };
-
-    let mut contents = String::new();
-    io::BufReader::new(file)
-        .read_to_string(&mut contents)
-        .map_err(|error| LocalCopyError::io("read filter file", path.to_path_buf(), error))?;
-
-    match options.parser() {
-        DirMergeParser::Whitespace { enforce_kind } => {
-            let enforce_kind = *enforce_kind;
-            let mut iter = contents.split_whitespace();
-            while let Some(token) = iter.next() {
-                if token.is_empty() {
-                    continue;
-                }
-
-                let token_lower = token.to_ascii_lowercase();
-                if token == "!" || token_lower == "clear" {
-                    if options.list_clear_allowed() {
-                        entries.rules.clear();
-                        entries.exclude_if_present.clear();
-                        continue;
-                    }
-                    let directive = if token == "!" { "!" } else { token };
-                    return Err(map_error(FilterParseError::new(format!(
-                        "list-clearing '{directive}' is not permitted in this filter file"
-                    ))));
-                }
-
-                if let Some(kind) = enforce_kind {
-                    let rule = match kind {
-                        DirMergeEnforcedKind::Include => FilterRule::include(token.to_string()),
-                        DirMergeEnforcedKind::Exclude => FilterRule::exclude(token.to_string()),
-                    };
-                    entries.push_rule(apply_dir_merge_rule_defaults(rule, options));
-                    continue;
-                }
-
-                let mut directive = token.to_string();
-                let lower = directive.to_ascii_lowercase();
-                let needs_argument = matches!(
-                    lower.as_str(),
-                    "merge"
-                        | "include"
-                        | "exclude"
-                        | "show"
-                        | "hide"
-                        | "protect"
-                        | "exclude-if-present"
-                ) || lower.starts_with("dir-merge");
-
-                if needs_argument {
-                    if let Some(next) = iter.next() {
-                        directive.push(' ');
-                        directive.push_str(next);
-                    }
-                }
-
-                match parse_filter_directive_line(&directive) {
-                    Ok(Some(ParsedFilterDirective::Rule(rule))) => {
-                        entries.push_rule(apply_dir_merge_rule_defaults(rule, options));
-                    }
-                    Ok(Some(ParsedFilterDirective::ExcludeIfPresent(rule))) => {
-                        entries.push_exclude_if_present(rule);
-                    }
-                    Ok(Some(ParsedFilterDirective::Clear)) => {
-                        entries.rules.clear();
-                        entries.exclude_if_present.clear();
-                    }
-                    Ok(Some(ParsedFilterDirective::Merge {
-                        path: merge_path,
-                        options: merge_options,
-                    })) => {
-                        let nested = if merge_path.is_absolute() {
-                            merge_path
-                        } else {
-                            let parent = path.parent().unwrap_or_else(|| Path::new("."));
-                            parent.join(merge_path)
-                        };
-                        if let Some(options_override) = merge_options {
-                            let nested_entries = load_dir_merge_rules_recursive(
-                                &nested,
-                                &options_override,
-                                visited,
-                            )?;
-                            entries.extend(nested_entries);
-                        } else {
-                            let nested_entries =
-                                load_dir_merge_rules_recursive(&nested, options, visited)?;
-                            entries.extend(nested_entries);
-                        }
-                    }
-                    Ok(None) => {}
-                    Err(error) => return Err(map_error(error)),
-                }
-            }
-        }
-        DirMergeParser::Lines {
-            enforce_kind,
-            allow_comments,
-        } => {
-            let enforce_kind = *enforce_kind;
-            let allow_comments = *allow_comments;
-            for line in contents.lines() {
-                let trimmed = line.trim();
-                if trimmed.is_empty() {
-                    continue;
-                }
-                if allow_comments && trimmed.starts_with('#') {
-                    continue;
-                }
-
-                if trimmed == "!" || trimmed.eq_ignore_ascii_case("clear") {
-                    if options.list_clear_allowed() {
-                        entries.rules.clear();
-                        entries.exclude_if_present.clear();
-                        continue;
-                    }
-                    return Err(map_error(FilterParseError::new(format!(
-                        "list-clearing '{}' is not permitted in this filter file",
-                        trimmed
-                    ))));
-                }
-
-                if let Some(kind) = enforce_kind {
-                    let rule = match kind {
-                        DirMergeEnforcedKind::Include => FilterRule::include(trimmed.to_string()),
-                        DirMergeEnforcedKind::Exclude => FilterRule::exclude(trimmed.to_string()),
-                    };
-                    entries.push_rule(apply_dir_merge_rule_defaults(rule, options));
-                    continue;
-                }
-
-                match parse_filter_directive_line(trimmed) {
-                    Ok(Some(ParsedFilterDirective::Rule(rule))) => {
-                        entries.push_rule(apply_dir_merge_rule_defaults(rule, options));
-                    }
-                    Ok(Some(ParsedFilterDirective::ExcludeIfPresent(rule))) => {
-                        entries.push_exclude_if_present(rule);
-                    }
-                    Ok(Some(ParsedFilterDirective::Merge {
-                        path: merge_path,
-                        options: merge_options,
-                    })) => {
-                        let nested = if merge_path.is_absolute() {
-                            merge_path
-                        } else {
-                            let parent = path.parent().unwrap_or_else(|| Path::new("."));
-                            parent.join(merge_path)
-                        };
-                        if let Some(options_override) = merge_options {
-                            let nested_entries = load_dir_merge_rules_recursive(
-                                &nested,
-                                &options_override,
-                                visited,
-                            )?;
-                            entries.extend(nested_entries);
-                        } else {
-                            let nested_entries =
-                                load_dir_merge_rules_recursive(&nested, options, visited)?;
-                            entries.extend(nested_entries);
-                        }
-                    }
-                    Ok(Some(ParsedFilterDirective::Clear)) => {
-                        entries.rules.clear();
-                        entries.exclude_if_present.clear();
-                    }
-                    Ok(None) => {}
-                    Err(error) => return Err(map_error(error)),
-                }
-            }
-        }
-    }
-
-    visited.pop();
-    Ok(entries)
-}
-
-#[derive(Debug)]
-enum ParsedFilterDirective {
-    Rule(FilterRule),
-    Merge {
-        path: PathBuf,
-        options: Option<DirMergeOptions>,
-    },
-    ExcludeIfPresent(ExcludeIfPresentRule),
-    Clear,
-}
-
-#[derive(Debug)]
-struct FilterParseError {
-    message: String,
-}
-
-impl FilterParseError {
-    fn new(message: impl Into<String>) -> Self {
-        Self {
-            message: message.into(),
-        }
-    }
-}
-
-impl fmt::Display for FilterParseError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.write_str(&self.message)
-    }
-}
-
-impl Error for FilterParseError {}
-
-fn parse_filter_directive_line(
-    text: &str,
-) -> Result<Option<ParsedFilterDirective>, FilterParseError> {
-    if text.is_empty() || text.starts_with('#') {
-        return Ok(None);
-    }
-
-    let trimmed = text.trim_start();
-    if trimmed.is_empty() {
-        return Ok(None);
-    }
-
-    let trimmed = trimmed.trim_end();
-
-    if trimmed == "!" || trimmed.eq_ignore_ascii_case("clear") {
-        return Ok(Some(ParsedFilterDirective::Clear));
-    }
-
-    if let Some(directive) = parse_short_merge_directive_line(trimmed)? {
-        return Ok(Some(directive));
-    }
-
-    if let Some(directive) = parse_merge_directive(trimmed)? {
-        return Ok(Some(directive));
-    }
-
-    if let Some(directive) = parse_dir_merge_directive(trimmed)? {
-        return Ok(Some(directive));
-    }
-
-    const EXCLUDE_IF_PRESENT_PREFIX: &str = "exclude-if-present";
-
-    if trimmed.len() >= EXCLUDE_IF_PRESENT_PREFIX.len()
-        && trimmed[..EXCLUDE_IF_PRESENT_PREFIX.len()]
-            .eq_ignore_ascii_case(EXCLUDE_IF_PRESENT_PREFIX)
-    {
-        let mut remainder = trimmed[EXCLUDE_IF_PRESENT_PREFIX.len()..]
-            .trim_start_matches(|ch: char| ch == '_' || ch.is_ascii_whitespace());
-        if let Some(rest) = remainder.strip_prefix('=') {
-            remainder = rest.trim_start_matches(|ch: char| ch == '_' || ch.is_ascii_whitespace());
-        }
-
-        let pattern_text = remainder.trim();
-        if pattern_text.is_empty() {
-            return Err(FilterParseError::new(
-                "filter directive 'exclude-if-present' requires a marker file",
-            ));
-        }
-
-        return Ok(Some(ParsedFilterDirective::ExcludeIfPresent(
-            ExcludeIfPresentRule::new(pattern_text),
-        )));
-    }
-
-    if let Some(remainder) = trimmed.strip_prefix('+') {
-        let pattern = remainder.trim_start();
-        if pattern.is_empty() {
-            return Err(FilterParseError::new("filter rule '+' requires a pattern"));
-        }
-        return Ok(Some(ParsedFilterDirective::Rule(FilterRule::include(
-            pattern.to_string(),
-        ))));
-    }
-
-    if let Some(remainder) = trimmed.strip_prefix('-') {
-        let pattern = remainder.trim_start();
-        if pattern.is_empty() {
-            return Err(FilterParseError::new("filter rule '-' requires a pattern"));
-        }
-        return Ok(Some(ParsedFilterDirective::Rule(FilterRule::exclude(
-            pattern.to_string(),
-        ))));
-    }
-
-    let mut parts = trimmed.splitn(2, char::is_whitespace);
-    let keyword = parts.next().unwrap_or("");
-    let remainder = parts.next().unwrap_or("").trim_start();
-
-    let handle_keyword = |pattern: &str,
-                          builder: fn(String) -> FilterRule|
-     -> Result<Option<ParsedFilterDirective>, FilterParseError> {
-        if pattern.is_empty() {
-            return Err(FilterParseError::new("filter directive missing pattern"));
-        }
-        Ok(Some(ParsedFilterDirective::Rule(builder(
-            pattern.to_string(),
-        ))))
-    };
-
-    if keyword.len() == 1 {
-        let shorthand = keyword.chars().next().unwrap().to_ascii_lowercase();
-        match shorthand {
-            'p' => {
-                return handle_keyword(remainder, FilterRule::protect);
-            }
-            'r' => {
-                return handle_keyword(remainder, FilterRule::risk);
-            }
-            's' => {
-                if remainder.is_empty() {
-                    return Err(FilterParseError::new("filter directive missing pattern"));
-                }
-                let rule = FilterRule::show(remainder.to_string());
-                return Ok(Some(ParsedFilterDirective::Rule(rule)));
-            }
-            'h' => {
-                if remainder.is_empty() {
-                    return Err(FilterParseError::new("filter directive missing pattern"));
-                }
-                let rule = FilterRule::hide(remainder.to_string());
-                return Ok(Some(ParsedFilterDirective::Rule(rule)));
-            }
-            _ => {}
-        }
-    }
-
-    if keyword.eq_ignore_ascii_case("include") {
-        return handle_keyword(remainder, FilterRule::include);
-    }
-
-    if keyword.eq_ignore_ascii_case("exclude") {
-        return handle_keyword(remainder, FilterRule::exclude);
-    }
-
-    if keyword.eq_ignore_ascii_case("show") {
-        if remainder.is_empty() {
-            return Err(FilterParseError::new("filter directive missing pattern"));
-        }
-        let rule = FilterRule::show(remainder.to_string());
-        return Ok(Some(ParsedFilterDirective::Rule(rule)));
-    }
-
-    if keyword.eq_ignore_ascii_case("hide") {
-        if remainder.is_empty() {
-            return Err(FilterParseError::new("filter directive missing pattern"));
-        }
-        let rule = FilterRule::hide(remainder.to_string());
-        return Ok(Some(ParsedFilterDirective::Rule(rule)));
-    }
-
-    if keyword.eq_ignore_ascii_case("protect") {
-        return handle_keyword(remainder, FilterRule::protect);
-    }
-
-    if keyword.eq_ignore_ascii_case("risk") {
-        return handle_keyword(remainder, FilterRule::risk);
-    }
-
-    Err(FilterParseError::new(format!(
-        "unsupported filter directive '{}'",
-        trimmed
-    )))
-}
-
-fn parse_merge_directive(text: &str) -> Result<Option<ParsedFilterDirective>, FilterParseError> {
-    const MERGE_PREFIX: &str = "merge";
-
-    if text.len() < MERGE_PREFIX.len() {
-        return Ok(None);
-    }
-
-    let (prefix, rest) = text.split_at(MERGE_PREFIX.len());
-    if !prefix.eq_ignore_ascii_case(MERGE_PREFIX) {
-        return Ok(None);
-    }
-
-    let mut remainder = rest.trim_start_matches(|ch: char| ch == '_' || ch.is_ascii_whitespace());
-    let mut modifiers = "";
-    if let Some(next) = remainder.strip_prefix(',') {
-        let mut split = next.splitn(2, |ch: char| ch.is_ascii_whitespace() || ch == '_');
-        modifiers = split.next().unwrap_or("");
-        remainder = split
-            .next()
-            .unwrap_or("")
-            .trim_start_matches(|ch: char| ch == '_' || ch.is_ascii_whitespace());
-    }
-
-    let (options, assume_cvsignore) = parse_merge_modifiers(modifiers, text, false)?;
-
-    if remainder == "-" {
-        return Err(FilterParseError::new(
-            "merge from standard input is not supported in .rsync-filter files",
-        ));
-    }
-
-    let path_text = remainder.trim_end();
-    let path_text = if path_text.is_empty() {
-        if assume_cvsignore {
-            ".cvsignore"
-        } else {
-            return Err(FilterParseError::new(
-                "merge directive requires a file path",
-            ));
-        }
-    } else {
-        path_text
-    };
-
-    let options = if modifiers.is_empty() && !assume_cvsignore {
-        None
-    } else {
-        Some(options)
-    };
-
-    Ok(Some(ParsedFilterDirective::Merge {
-        path: PathBuf::from(path_text),
-        options,
-    }))
-}
-
-fn parse_short_merge_directive_line(
-    text: &str,
-) -> Result<Option<ParsedFilterDirective>, FilterParseError> {
-    let mut chars = text.chars();
-    let first = match chars.next() {
-        Some(first) => first,
-        None => return Ok(None),
-    };
-
-    let allow_extended = match first {
-        '.' => false,
-        ':' => true,
-        _ => return Ok(None),
-    };
-
-    let remainder = chars.as_str();
-    let (modifiers, rest) = split_short_rule_modifiers(remainder);
-    let (options, assume_cvsignore) = parse_merge_modifiers(modifiers, text, allow_extended)?;
-
-    let pattern = rest.trim();
-    let pattern = if pattern.is_empty() {
-        if assume_cvsignore {
-            ".cvsignore"
-        } else if allow_extended {
-            return Err(FilterParseError::new(format!(
-                "dir-merge directive '{}' is missing a file name",
-                text
-            )));
-        } else {
-            return Err(FilterParseError::new(format!(
-                "merge directive '{}' is missing a file path",
-                text
-            )));
-        }
-    } else {
-        pattern
-    };
-
-    if allow_extended {
-        return Ok(Some(ParsedFilterDirective::Merge {
-            path: PathBuf::from(pattern),
-            options: Some(options),
-        }));
-    }
-
-    let options = if modifiers.is_empty() && !assume_cvsignore {
-        None
-    } else {
-        Some(options)
-    };
-
-    Ok(Some(ParsedFilterDirective::Merge {
-        path: PathBuf::from(pattern),
-        options,
-    }))
-}
-
-fn split_short_rule_modifiers(text: &str) -> (&str, &str) {
-    if text.is_empty() {
-        return ("", "");
-    }
-
-    if let Some(rest) = text.strip_prefix(',') {
-        let mut parts = rest.splitn(2, |ch: char| ch.is_ascii_whitespace() || ch == '_');
-        let modifiers = parts.next().unwrap_or("");
-        let remainder = parts.next().unwrap_or("");
-        let remainder =
-            remainder.trim_start_matches(|ch: char| ch == '_' || ch.is_ascii_whitespace());
-        return (modifiers, remainder);
-    }
-
-    let mut chars = text.chars();
-    match chars.next() {
-        None => ("", ""),
-        Some(first) if first.is_ascii_whitespace() || first == '_' => {
-            let remainder =
-                text.trim_start_matches(|ch: char| ch == '_' || ch.is_ascii_whitespace());
-            ("", remainder)
-        }
-        Some(_) => {
-            let mut len = 0;
-            for ch in text.chars() {
-                if ch.is_ascii_whitespace() || ch == '_' {
-                    break;
-                }
-                len += ch.len_utf8();
-            }
-            let modifiers = &text[..len];
-            let remainder =
-                text[len..].trim_start_matches(|ch: char| ch == '_' || ch.is_ascii_whitespace());
-            (modifiers, remainder)
-        }
-    }
-}
-
-fn parse_merge_modifiers(
-    modifiers: &str,
-    directive: &str,
-    allow_extended: bool,
-) -> Result<(DirMergeOptions, bool), FilterParseError> {
-    let label = if allow_extended { "dir-merge" } else { "merge" };
-    let mut options = if allow_extended {
-        DirMergeOptions::default()
-    } else {
-        DirMergeOptions::default().allow_list_clearing(true)
-    };
-    let mut enforced: Option<DirMergeEnforcedKind> = None;
-    let mut saw_include = false;
-    let mut saw_exclude = false;
-    let mut assume_cvsignore = false;
-
-    for modifier in modifiers.chars() {
-        let lower = modifier.to_ascii_lowercase();
-        match lower {
-            '-' => {
-                if saw_include {
-                    let message = format!(
-                        "{label} directive '{}' cannot combine '+' and '-' modifiers",
-                        directive
-                    );
-
-                    return Err(FilterParseError::new(message));
-                }
-                saw_exclude = true;
-                enforced = Some(DirMergeEnforcedKind::Exclude);
-            }
-            '+' => {
-                if saw_exclude {
-                    let message = format!(
-                        "{label} directive '{}' cannot combine '+' and '-' modifiers",
-                        directive
-                    );
-                    return Err(FilterParseError::new(message));
-                }
-                saw_include = true;
-                enforced = Some(DirMergeEnforcedKind::Include);
-            }
-            'c' => {
-                if saw_include {
-                    let message = format!(
-                        "{label} directive '{}' cannot combine 'C' with '+' or '-'",
-                        directive
-                    );
-                    return Err(FilterParseError::new(message));
-                }
-                saw_exclude = true;
-                enforced = Some(DirMergeEnforcedKind::Exclude);
-                options = options
-                    .use_whitespace()
-                    .allow_comments(false)
-                    .allow_list_clearing(true)
-                    .inherit(false);
-                assume_cvsignore = true;
-            }
-            'e' => {
-                if allow_extended {
-                    options = options.exclude_filter_file(true);
-                } else {
-                    let message = format!(
-                        "merge directive '{}' uses unsupported modifier '{}'",
-                        directive, modifier
-                    );
-                    return Err(FilterParseError::new(message));
-                }
-            }
-            'n' => {
-                if allow_extended {
-                    options = options.inherit(false);
-                } else {
-                    let message = format!(
-                        "merge directive '{}' uses unsupported modifier '{}'",
-                        directive, modifier
-                    );
-                    return Err(FilterParseError::new(message));
-                }
-            }
-            'w' => {
-                if allow_extended {
-                    options = options.use_whitespace().allow_comments(false);
-                } else {
-                    let message = format!(
-                        "merge directive '{}' uses unsupported modifier '{}'",
-                        directive, modifier
-                    );
-                    return Err(FilterParseError::new(message));
-                }
-            }
-            's' => {
-                if allow_extended {
-                    options = options.sender_modifier();
-                } else {
-                    let message = format!(
-                        "merge directive '{}' uses unsupported modifier '{}'",
-                        directive, modifier
-                    );
-                    return Err(FilterParseError::new(message));
-                }
-            }
-            'r' => {
-                if allow_extended {
-                    options = options.receiver_modifier();
-                } else {
-                    let message = format!(
-                        "merge directive '{}' uses unsupported modifier '{}'",
-                        directive, modifier
-                    );
-                    return Err(FilterParseError::new(message));
-                }
-            }
-            '/' => {
-                if allow_extended {
-                    options = options.anchor_root(true);
-                } else {
-                    let message = format!(
-                        "merge directive '{}' uses unsupported modifier '{}'",
-                        directive, modifier
-                    );
-                    return Err(FilterParseError::new(message));
-                }
-            }
-            _ => {
-                let message = format!(
-                    "{label} directive '{}' uses unsupported modifier '{}'",
-                    directive, modifier
-                );
-                return Err(FilterParseError::new(message));
-            }
-        }
-    }
-
-    options = options.with_enforced_kind(enforced);
-    if !allow_extended && !options.list_clear_allowed() {
-        options = options.allow_list_clearing(true);
-    }
-
-    Ok((options, assume_cvsignore))
-}
-
-fn parse_dir_merge_directive(
-    text: &str,
-) -> Result<Option<ParsedFilterDirective>, FilterParseError> {
-    const DIR_MERGE_PREFIX: &str = "dir-merge";
-
-    if text.len() < DIR_MERGE_PREFIX.len() {
-        return Ok(None);
-    }
-
-    let (prefix, mut remainder) = text.split_at(DIR_MERGE_PREFIX.len());
-    if !prefix.eq_ignore_ascii_case(DIR_MERGE_PREFIX) {
-        return Ok(None);
-    }
-
-    if let Some(ch) = remainder.chars().next() {
-        if ch != ',' && !ch.is_ascii_whitespace() {
-            return Ok(None);
-        }
-    }
-
-    remainder = remainder.trim_start();
-
-    let mut modifiers = "";
-    if let Some(rest) = remainder.strip_prefix(',') {
-        let mut split = rest.splitn(2, char::is_whitespace);
-        modifiers = split.next().unwrap_or("");
-        remainder = split.next().unwrap_or("").trim_start();
-    }
-
-    let mut options = DirMergeOptions::default();
-    let mut saw_plus = false;
-    let mut saw_minus = false;
-    let mut used_cvs_default = false;
-
-    for modifier in modifiers.chars() {
-        let lower = modifier.to_ascii_lowercase();
-        match lower {
-            '-' => {
-                if saw_plus {
-                    let message = format!(
-                        "dir-merge directive '{}' cannot combine '+' and '-' modifiers",
-                        text
-                    );
-                    return Err(FilterParseError::new(message));
-                }
-                saw_minus = true;
-                options = options.with_enforced_kind(Some(DirMergeEnforcedKind::Exclude));
-            }
-            '+' => {
-                if saw_minus {
-                    let message = format!(
-                        "dir-merge directive '{}' cannot combine '+' and '-' modifiers",
-                        text
-                    );
-                    return Err(FilterParseError::new(message));
-                }
-                saw_plus = true;
-                options = options.with_enforced_kind(Some(DirMergeEnforcedKind::Include));
-            }
-            'n' => {
-                options = options.inherit(false);
-            }
-            'e' => {
-                options = options.exclude_filter_file(true);
-            }
-            'w' => {
-                options = options.use_whitespace();
-                options = options.allow_comments(false);
-            }
-            's' => {
-                options = options.sender_modifier();
-            }
-            'r' => {
-                options = options.receiver_modifier();
-            }
-            '/' => {
-                options = options.anchor_root(true);
-            }
-            'c' => {
-                used_cvs_default = true;
-                options = options.with_enforced_kind(Some(DirMergeEnforcedKind::Exclude));
-                options = options.use_whitespace();
-                options = options.allow_comments(false);
-                options = options.inherit(false);
-                options = options.allow_list_clearing(true);
-            }
-            _ => {
-                let message = format!(
-                    "dir-merge directive '{}' uses unsupported modifier '{}'",
-                    text, modifier
-                );
-                return Err(FilterParseError::new(message));
-            }
-        }
-    }
-
-    let path_text = if remainder.is_empty() {
-        if used_cvs_default {
-            ".cvsignore"
-        } else {
-            let message = format!("dir-merge directive '{}' is missing a file name", text);
-            return Err(FilterParseError::new(message));
-        }
-    } else {
-        remainder
-    };
-
-    Ok(Some(ParsedFilterDirective::Merge {
-        path: PathBuf::from(path_text),
-        options: Some(options),
-    }))
-}
-
 #[cfg(feature = "xattr")]
 fn sync_xattrs_if_requested(
     preserve_xattrs: bool,
@@ -2685,227 +1839,6 @@ fn sync_acls_if_requested(
         sync_acls(source, destination, follow_symlinks).map_err(map_metadata_error)?;
     }
     Ok(())
-}
-
-/// Error produced when planning or executing a local copy fails.
-#[derive(Debug)]
-pub struct LocalCopyError {
-    kind: LocalCopyErrorKind,
-}
-
-impl LocalCopyError {
-    fn new(kind: LocalCopyErrorKind) -> Self {
-        Self { kind }
-    }
-
-    /// Constructs an error representing missing operands.
-    #[must_use]
-    pub fn missing_operands() -> Self {
-        Self::new(LocalCopyErrorKind::MissingSourceOperands)
-    }
-
-    /// Constructs an invalid-argument error.
-    #[must_use]
-    pub fn invalid_argument(reason: LocalCopyArgumentError) -> Self {
-        Self::new(LocalCopyErrorKind::InvalidArgument(reason))
-    }
-
-    /// Constructs an error indicating that the deletion limit was exceeded.
-    #[must_use]
-    pub fn delete_limit_exceeded(skipped: u64) -> Self {
-        Self::new(LocalCopyErrorKind::DeleteLimitExceeded { skipped })
-    }
-
-    /// Constructs an I/O error with action context.
-    #[must_use]
-    pub fn io(action: &'static str, path: PathBuf, source: io::Error) -> Self {
-        Self::new(LocalCopyErrorKind::Io {
-            action,
-            path,
-            source,
-        })
-    }
-
-    /// Constructs an error representing an inactivity timeout.
-    #[must_use]
-    pub fn timeout(duration: Duration) -> Self {
-        Self::new(LocalCopyErrorKind::Timeout { duration })
-    }
-
-    /// Returns the exit code that mirrors upstream rsync's behaviour.
-    #[must_use]
-    pub const fn exit_code(&self) -> i32 {
-        match self.kind {
-            LocalCopyErrorKind::MissingSourceOperands => MISSING_OPERANDS_EXIT_CODE,
-            LocalCopyErrorKind::InvalidArgument(_) | LocalCopyErrorKind::Io { .. } => {
-                INVALID_OPERAND_EXIT_CODE
-            }
-            LocalCopyErrorKind::Timeout { .. } => TIMEOUT_EXIT_CODE,
-            LocalCopyErrorKind::DeleteLimitExceeded { .. } => MAX_DELETE_EXIT_CODE,
-        }
-    }
-
-    /// Provides access to the underlying error kind.
-    #[must_use]
-    pub fn kind(&self) -> &LocalCopyErrorKind {
-        &self.kind
-    }
-
-    /// Consumes the error and returns its kind.
-    #[must_use]
-    pub fn into_kind(self) -> LocalCopyErrorKind {
-        self.kind
-    }
-}
-
-impl fmt::Display for LocalCopyError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match &self.kind {
-            LocalCopyErrorKind::MissingSourceOperands => {
-                write!(
-                    f,
-                    "missing source operands: supply at least one source and a destination"
-                )
-            }
-            LocalCopyErrorKind::InvalidArgument(reason) => write!(f, "{}", reason.message()),
-            LocalCopyErrorKind::Io {
-                action,
-                path,
-                source,
-            } => {
-                write!(f, "failed to {action} '{}': {source}", path.display())
-            }
-            LocalCopyErrorKind::Timeout { duration } => {
-                write!(
-                    f,
-                    "transfer timed out after {:.3} seconds without progress",
-                    duration.as_secs_f64()
-                )
-            }
-            LocalCopyErrorKind::DeleteLimitExceeded { skipped } => {
-                let noun = if *skipped == 1 { "entry" } else { "entries" };
-                write!(
-                    f,
-                    "Deletions stopped due to --max-delete limit ({} {noun} skipped)",
-                    skipped
-                )
-            }
-        }
-    }
-}
-
-impl Error for LocalCopyError {
-    fn source(&self) -> Option<&(dyn Error + 'static)> {
-        match &self.kind {
-            LocalCopyErrorKind::Io { source, .. } => Some(source),
-            _ => None,
-        }
-    }
-}
-
-/// Classification of local copy failures.
-#[derive(Debug)]
-pub enum LocalCopyErrorKind {
-    /// No operands were supplied.
-    MissingSourceOperands,
-    /// Operands were invalid.
-    InvalidArgument(LocalCopyArgumentError),
-    /// Filesystem interaction failed.
-    Io {
-        /// Action being performed.
-        action: &'static str,
-        /// Path involved in the failure.
-        path: PathBuf,
-        /// Underlying error.
-        source: io::Error,
-    },
-    /// The transfer exceeded the configured inactivity timeout.
-    Timeout {
-        /// Duration of inactivity that triggered the timeout.
-        duration: Duration,
-    },
-    /// Deletions were halted because the configured limit was exceeded.
-    DeleteLimitExceeded {
-        /// Number of entries that were skipped after reaching the limit.
-        skipped: u64,
-    },
-}
-
-impl LocalCopyErrorKind {
-    /// Returns the action, path, and source error for [`LocalCopyErrorKind::Io`] values.
-    #[must_use]
-    pub fn as_io(&self) -> Option<(&'static str, &Path, &io::Error)> {
-        match self {
-            Self::Io {
-                action,
-                path,
-                source,
-            } => Some((action, path.as_path(), source)),
-            Self::Timeout { .. } => None,
-            _ => None,
-        }
-    }
-}
-
-/// Detailed reason for operand validation failures.
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub enum LocalCopyArgumentError {
-    /// A source operand was empty.
-    EmptySourceOperand,
-    /// The destination operand was empty.
-    EmptyDestinationOperand,
-    /// Multiple sources targeted a non-directory destination.
-    DestinationMustBeDirectory,
-    /// Unable to determine the directory name from the source operand.
-    DirectoryNameUnavailable,
-    /// Unable to determine the file name from the source operand.
-    FileNameUnavailable,
-    /// Unable to determine the link name from the source operand.
-    LinkNameUnavailable,
-    /// Encountered a file type that is unsupported.
-    UnsupportedFileType,
-    /// Attempted to replace an existing directory with a symbolic link.
-    ReplaceDirectoryWithSymlink,
-    /// Attempted to replace an existing directory with a regular file.
-    ReplaceDirectoryWithFile,
-    /// Attempted to replace an existing directory with a special file.
-    ReplaceDirectoryWithSpecial,
-    /// Attempted to replace a non-directory with a directory.
-    ReplaceNonDirectoryWithDirectory,
-    /// Encountered an operand that refers to a remote host or module.
-    RemoteOperandUnsupported,
-}
-
-impl LocalCopyArgumentError {
-    /// Returns the canonical diagnostic message associated with the error.
-    #[must_use]
-    pub const fn message(self) -> &'static str {
-        match self {
-            Self::EmptySourceOperand => "source operands must be non-empty",
-            Self::EmptyDestinationOperand => "destination operand must be non-empty",
-            Self::DestinationMustBeDirectory => {
-                "destination must be an existing directory when copying multiple sources"
-            }
-            Self::DirectoryNameUnavailable => "cannot determine directory name",
-            Self::FileNameUnavailable => "cannot determine file name",
-            Self::LinkNameUnavailable => "cannot determine link name",
-            Self::UnsupportedFileType => "unsupported file type encountered",
-            Self::ReplaceDirectoryWithSymlink => {
-                "cannot replace existing directory with symbolic link"
-            }
-            Self::ReplaceDirectoryWithFile => "cannot replace existing directory with regular file",
-            Self::ReplaceDirectoryWithSpecial => {
-                "cannot replace existing directory with special file"
-            }
-            Self::ReplaceNonDirectoryWithDirectory => {
-                "cannot replace non-directory destination with directory"
-            }
-            Self::RemoteOperandUnsupported => concat!(
-                "remote operands are not supported: this build handles local filesystem copies only; ",
-                "set OC_RSYNC_FALLBACK to point to an upstream rsync binary for remote transfers",
-            ),
-        }
-    }
 }
 
 /// Source operand within a [`LocalCopyPlan`].
@@ -6464,5 +5397,5 @@ fn create_symlink(target: &Path, source: &Path, destination: &Path) -> io::Resul
 mod tests;
 
 #[cfg(test)]
-#[path = "local_copy/filter_program_internal_tests.rs"]
+#[path = "filter_program_internal_tests.rs"]
 mod filter_program_internal_tests;
